Move stub parser to its own file to satisfy PHPCS one-class-per-file rule
